feat: mock in conftest
